% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data.tax_slabs}
\alias{data.tax_slabs}
\title{tax slabs with updated tax rate}
\format{
An object of class \code{data.frame} with 6 rows and 3 columns.
}
\usage{
data.tax_slabs
}
\description{
tax slabs with updated tax rate
}
\keyword{datasets}
