% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax_calculator.R
\name{taxCalculator}
\alias{taxCalculator}
\title{tax calculator}
\usage{
taxCalculator(
  x,
  sec_80c = 0,
  standard_deduction = 50000,
  no_tax_bracket = 250000
)
}
\arguments{
\item{x}{annual income}

\item{sec_80c}{investment declaration under section 80C}

\item{standard_deduction}{constant, could change over time though}

\item{no_tax_bracket}{non taxable bracket, could change over time though.}
}
\description{
Calculation of tax amount in India
}
\examples{
taxCalculator(300000)
taxCalculator(600000)
taxCalculator(1000000)
taxCalculator(2000000, sec_80c = 120000)

}
